requirements.txt
asgiref~=3.3.1
dj-database-url~=0.5.0
Django~=3.1.6
django-heroku~=0.3.1
django-summernote~=0.8.11.6
gunicorn~=20.0.4
Pillow~=8.1.0
psycopg2~=2.8.6
pytz~=2021.1
sqlparse~=0.4.1
whitenoise~=5.2.0
cloudinary~=1.24.0
django-taggit~=1.3.0
django-ckeditor~=6.0.0
jwplatform~=2.1.1